Partial implementation of units tests for WeightAndBlend with spot data.
